requirements_complete.txt
# StreamForge-Pro Complete Requirements
# Core dependencies
yt-dlp>=2023.12.30
ffmpeg-python>=0.2.0
requests>=2.31.0
aiohttp>=3.9.0

# Database
aiosqlite>=0.19.0
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI Interface
click>=8.1.0
rich>=13.7.0
prompt-toolkit>=3.0.0

# GUI Interface (optional)
PyQt6>=6.6.0
qasync>=0.24.0

# Web Interface
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
jinja2>=3.1.0

# Telegram Bot
python-telegram-bot>=20.0

# Cloud Upload
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
dropbox>=11.0.0

# Utilities
pillow>=10.1.0
psutil>=5.9.0
tqdm>=4.66.0

# Security
cryptography>=41.0.0
pyjwt>=2.8.0

# Testing (dev)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
